refactor: 공통 style 분리
