+++ src/components/org/OrgAdmin/CommonSection/GenerationInformation.tsx
@@ -9,7 +9,7 @@ import {
   StTitle,
   StTitleWrapper,
   StWrapper,
-} from './style';
+} from '../style';
 
 const GenerationInformation = () => {
   const {
